trello refactor
