+++ cmd/trellogithub/main.go
@@ -3,33 +3,33 @@ package main
 import (
 	"log"
 
-	"github.com/merico-dev/stream/internal/pkg/plugin/githubintegrations"
+	"github.com/merico-dev/stream/internal/pkg/plugin/trellogithub"
 )
 
 // NAME is the name of this DevStream plugin.
-const NAME = "githubintegrations"
+const NAME = "trellogithub"
 
 // Plugin is the type used by DevStream core. It's a string.
 type Plugin string
 
 // Install implements the installation of some GitHub Actions workflows.
 func (p Plugin) Install(options *map[string]interface{}) (bool, error) {
-	return githubintegrations.Install(options)
+	return trellogithub.Install(options)
 }
 
 // Reinstall implements the installation of some GitHub Actions workflows.
 func (p Plugin) Reinstall(options *map[string]interface{}) (bool, error) {
-	return githubintegrations.Reinstall(options)
+	return trellogithub.Reinstall(options)
 }
 
 // Uninstall implements the installation of some GitHub Actions workflows.
 func (p Plugin) Uninstall(options *map[string]interface{}) (bool, error) {
-	return githubintegrations.Uninstall(options)
+	return trellogithub.Uninstall(options)
 }
 
 // IsHealthy implements the healthy check of GitHub Actions workflows.
 func (p Plugin) IsHealthy(options *map[string]interface{}) (bool, error) {
-	return githubintegrations.IsHealthy(options)
+	return trellogithub.IsHealthy(options)
 }
 
 // DevStreamPlugin is the exported variable used by the DevStream core.
