+++ internal/pkg/plugin/trellogithub/jobs.go
@@ -1,4 +1,4 @@
-package githubintegrations
+package trellogithub
 
 // Jobs is the struct for github actions custom config.
 type Jobs struct {
