+++ internal/pkg/plugin/trellogithub/trellogithub_suit_test.go
@@ -1,4 +1,4 @@
-package githubintegrations
+package trellogithub
 
 import (
 	"testing"
