+++ internal/pkg/plugin/trellogithub/trellogithub_test.go
@@ -1,4 +1,4 @@
-package githubintegrations
+package trellogithub
 
 import (
 	. "github.com/onsi/ginkgo/v2"
